Support configuration against protobuf 3.11 on Windows
Before this modification, the use of the HINTS parameter in find_package
make the CMake use the installed protobuf-config.cmake instead of
the CMake's provided FindProtobuf.cmake that was used on non-WIN32.
Unfortunatly, the variables and macros exported by the protobuf 3.11's
protobuf-config.cmake are not compatible with the one used in the project,
so also on Windows we use the default signature of find_package.